fix: removed unnecessary type conversion

+++ Backend/internal/server/middleware/auth.go
@@ -26,7 +26,7 @@ func (m *Middlewares) Auth() echo.MiddlewareFunc {
 			}
 
 			user := domain.User{}
-			if err := m.repos.User.GetByID(uint64(claims.ID), &user); err != nil {
+			if err := m.repos.User.GetByID(claims.ID, &user); err != nil {
 				return echo.NewHTTPError(401, "Unauthorized, failed to get user")
 			}
 			user.Password = "********"
